requirements.txt
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coverage==6.2
Deprecated==1.2.13
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongomock==3.23.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pymongo==4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-bsonjs==0.2.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.0
sentinels==1.0.0
six==1.16.0
toml==0.10.2
tomli==1.2.2
twilio==7.4.0
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
